requirements.txt
aiogram==3.13.1
httpx==0.27.2
requests==2.32.3
fake-useragent==1.5.1
aiofiles==24.1.0
python-dotenv==1.0.1
